refactor: rename assets
